Find K-th smallest number from a number n
Given a non-negative number n and a value k. Find the k-th smallest number that can be formed using the digits of the given number n. It is guaranteed that the k-th smallest number can be formed. Note that the number could be very large and may not even fit into long long int.